reorganize frontend to make merging easier
