+++ frontend/src/pages/Game.tsx
